Bikeshare data analysis
This is a python script that pulls Bikeshare data from three CSV files stored on my google drive account for three major cities: Chicago, New York and Washington, then runs some simple statistical analysis to gain insights into the data. The script is interactive and will be requesting input like city name, month and day of the week to display the related insights. Alternatively, it is also possible to select a city then display insights for all months and days of the week.
It is also possible to keep looking at insights for different combinations of input until one decides to tell the script to terminate by answering "No" to the final question.

The links for the data files are in the script.