Bikeshare data analysis
This is a python script that pulls Bikeshare data from three CSV files stored on my google drive for three major cities: Chicago, New York and Washington. The script then runs some simple statistical analysis on the data generating insights. The script is also interactive, meaning it will request input like city name, month and day of the week in order to return insights related to that input specifically. Alternatively, once a city is selected. all the information related to all months and all days of the week can be considered in the analysis. Also, sample from the raw data can be displayed and the script will keep running to allow looking at insights for different input combinations until one decides to terminate the script by answering "No" to the last question the script will ask.

The links for the data files used are in the script.